broadband-modem-qmi: Fix compilation error due to variable type change
otherwise we fail with:

CC       ModemManager-mm-broadband-modem-qmi.o
mm-broadband-modem-qmi.c: In function ‘get_signal_info_ready’:
mm-broadband-modem-qmi.c:2309:5: error: passing argument 3 of ‘signal_info_get_quality’ from incompatible pointer type [-Werror]
     if (!signal_info_get_quality (ctx->self, output, &quality)) {
     ^
mm-broadband-modem-qmi.c:2236:1: note: expected ‘gint8 *’ but argument is of type ‘guint *’
 signal_info_get_quality (MMBroadbandModemQmi *self,
 ^
cc1: all warnings being treated as errors

+++ src/mm-broadband-modem-qmi.c
@@ -2290,7 +2290,7 @@ get_signal_info_ready (QmiClientNas *client,
 {
     QmiMessageNasGetSignalInfoOutput *output;
     GError *error = NULL;
-    guint quality = 0;
+    gint8 quality = 0;
 
     output = qmi_client_nas_get_signal_info_finish (client, res, &error);
     if (!output) {
